@@ -14,22 +14,22 @@
     "anhskohbo/no-captcha": "^3.0",
     "aws/aws-sdk-php": "~3.0",
     "axn/laravel-pk-int-to-bigint": "^2.1",
-    "cviebrock/eloquent-sluggable": "^9.0",
+    "cviebrock/eloquent-sluggable": "^10.0",
     "czim/laravel-paperclip": "^5.0",
-    "graham-campbell/markdown": "14.0.x-dev",
+    "doctrine/dbal": "^3.0",
+    "graham-campbell/markdown": "^15.0",
     "guzzlehttp/guzzle": "^7.0.1",
-    "jenssegers/mongodb": "^3.8",
-    "joggapp/laravel-aws-sns": "^5.1",
     "laracasts/utilities": "^3.2",
-    "laravel/framework": "^9.0",
+    "laravel/framework": "^10.0",
     "laravel/nova": "~4.0",
-    "laravel/sanctum": "^2.8",
+    "laravel/sanctum": "^3.2",
     "laravel/tinker": "^2.5",
-    "laravel/ui": "^3.0",
+    "laravel/ui": "^4.0",
     "league/csv": "^9.6",
     "league/flysystem-aws-s3-v3": "^3.0",
     "league/fractal": "^0.20",
     "league/oauth2-client": "^2.6",
+    "mongodb/laravel-mongodb": "^4.0@RC",
     "neitanod/forceutf8": "~2.0",
     "outl1ne/nova-sortable": "^3.4",
     "pda/pheanstalk": "^4.0",
@@ -46,11 +46,10 @@
     "barryvdh/laravel-ide-helper": "^2.8",
     "beyondcode/laravel-dump-server": "^1.4",
     "fakerphp/faker": "^1.9.1",
-    "kitloong/laravel-migrations-generator": "^v5.2.0",
     "mockery/mockery": "^1.4.2",
-    "nunomaduro/collision": "^6.1",
-    "phpunit/phpunit": "^9.3.3",
-    "spatie/laravel-ignition": "^1.0"
+    "nunomaduro/collision": "^7.0",
+    "phpunit/phpunit": "^10.0",
+    "spatie/laravel-ignition": "^2.0"
   },
   "repositories": {
     "nova": {
@@ -80,7 +79,7 @@
       "Tests\\": "tests/"
     }
   },
-  "minimum-stability": "dev",
+  "minimum-stability": "stable",
   "prefer-stable": true,
   "scripts": {
     "post-autoload-dump": [
