@@ -19,7 +19,7 @@
 
 namespace App\Models;
 
-use Jenssegers\Mongodb\Eloquent\Model;
+use MongoDB\Laravel\Eloquent\Model;
 
 /**
  * Class BaseMongoModel
@@ -48,20 +48,4 @@ class BaseMongoModel extends Model
      * @inheritDoc
      */
     public $incrementing = false;
-
-    /**
-     * Enable casts in models.
-     *
-     * @param string $key
-     * @param mixed $value
-     * @return \Jenssegers\Mongodb\Eloquent\Model|mixed|void
-     */
-    public function setAttribute($key, $value)
-    {
-        if ($this->hasCast($key)) {
-            $value = $this->castAttribute($key, $value);
-        }
-
-        return parent::setAttribute($key, $value);
-    }
 }
