@@ -19,7 +19,7 @@
 
 namespace App\Models;
 
-use Jenssegers\Mongodb\Eloquent\HybridRelations;
+use MongoDB\Laravel\Eloquent\HybridRelations;
 
 /**
  * Class EventTranscription
@@ -58,18 +58,26 @@ class EventTranscription extends BaseEloquentModel
      * @var array
      */
     protected $casts = [
-        'classification_id' => 'int',
-        'event_id' => 'int',
-        'team_id' => 'int',
-        'user_id' => 'int',
+        'classification_id' => 'integer',
+        'event_id' => 'integer',
+        'team_id' => 'integer',
+        'user_id' => 'integer',
     ];
+
+    /**
+     * Model Boot
+     */
+    public static function boot()
+    {
+        parent::boot();
+    }
 
     /**
      * Transcription relationship.
      *
      * @return \Illuminate\Database\Eloquent\Relations\HasOne
      */
-    public function transcription()
+    public function transcription(): \Illuminate\Database\Eloquent\Relations\HasOne
     {
         return $this->hasOne(PanoptesTranscription::class,'_id', 'classification_id');
     }
@@ -79,7 +87,7 @@ class EventTranscription extends BaseEloquentModel
      *
      * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
      */
-    public function event()
+    public function event(): \Illuminate\Database\Eloquent\Relations\BelongsTo
     {
         return $this->belongsTo(Event::class);
     }
@@ -89,7 +97,7 @@ class EventTranscription extends BaseEloquentModel
      *
      * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
      */
-    public function team()
+    public function team(): \Illuminate\Database\Eloquent\Relations\BelongsTo
     {
         return $this->belongsTo(EventTeam::class);
     }
@@ -99,7 +107,7 @@ class EventTranscription extends BaseEloquentModel
      *
      * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
      */
-    public function user()
+    public function user(): \Illuminate\Database\Eloquent\Relations\BelongsTo
     {
         return $this->belongsTo(EventUser::class);
     }
