@@ -19,6 +19,7 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Support\Carbon;
 use MongoDB\BSON\UTCDateTime;
 
@@ -39,12 +40,14 @@ class PanoptesTranscription extends BaseMongoModel
      * @var string[]
      */
     protected $casts = [
-        'subject_id' => 'int',
-        'classification_id' => 'int',
-        'workflow_id' => 'int',
-        'subject_expeditionId' => 'int',
-        'subject_projectId' => 'int',
-        'transcription_id' => 'string'
+        'subject_id' => 'integer',
+        'classification_id' => 'integer',
+        'workflow_id' => 'integer',
+        'subject_expeditionId' => 'integer',
+        'subject_projectId' => 'integer',
+        'transcription_id' => 'string',
+        //'classification_finished_at' => 'datetime',
+        //'classification_started_at' => 'datetime' TODO check datetime casts work with mongodb
     ];
 
     /**
@@ -53,11 +56,6 @@ class PanoptesTranscription extends BaseMongoModel
      * @var array
      */
     protected $orderBy = [[]];
-
-    protected static function boot()
-    {
-        parent::boot();
-    }
 
     /**
      * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
@@ -110,5 +108,17 @@ class PanoptesTranscription extends BaseMongoModel
     public function setClassificationStartedAtAttribute($value)
     {
         $this->attributes['classification_started_at'] = new UTCDateTime(Carbon::parse($value)->timestamp * 1000);
+    }
+
+    /**
+     * New method for getting and setting attributes.
+     * TODO Change attribute setters and mutators.
+     */
+    protected function firstName(): Attribute
+    {
+        return Attribute::make(
+            get: fn (string $value) => ucfirst($value),
+            set: fn (string $value) => strtolower($value),
+        );
     }
 }
