@@ -21,7 +21,7 @@ namespace App\Models;
 
 use App\Facades\DateHelper;
 use App\Presenters\ProjectPresenter;
-use Jenssegers\Mongodb\Eloquent\HybridRelations;
+use MongoDB\Laravel\Eloquent\HybridRelations;
 use Illuminate\Support\Facades\Config;
 use Czim\Paperclip\Contracts\AttachableInterface;
 use Czim\Paperclip\Model\PaperclipTrait;
